requirements.txt
streamlit
python-docx
google-generativeai
python-dotenv
